wificond: Add interface event callback to notify number of connected clients
Currently when Softap is enabled and a client connects to ApInterface,
OnStationEvent is received internally to wificond. This data can be
queried by wifi service / softap manager to know about number of connected
clients.

Enhance this via a callback mechanism which can send connected/disconnect
client information as soon as OnStationEvent is received, along with
mac_address of client

Change-Id: I86cdb0c1d0e7ac38a5a6b600ab92b63e91a9ce3c
CRs-Fixed: 2072184