wificond: Enable support for STA+SAP Concurrency.
This introduces a new param new_sap_interface which would be set
to true on creation of new virtual interface for SoftAp. If set,
this would help in identifying softap0 interface while creating
SoftAP interface.

Additionally this commit adds AP_MODE to InterfaceMode which help
distinguishing which interface should be used for SoftAP/STA in
SetUpInterfaces.

CRs-Fixed: 2069861
Change-Id: I64cd5999446b7be7bef71be495aaec42087815ee