wifi: Avoid Change interface from SAP to STA during SAP stop.
During SAP OFF, wificond sets the interface mode back to station mode
(from the AP mode) after stopping hostapd. This would not be required for a
STA + SAP concurrent scenario as changing an interface to STA would assign all
the driver resources to the respective adapter. Further, added interface has to
be removed on a SAP OFF. Also, mode change from wificond is not required for
the standalone SAP mode as the wpa_supplicant/hostapd initialization
(on a STA/SAP ON)will set the respective mode (STA/SAP mode) to the adapter.
Hence avoid the change interface.
CRs-Fixed: 2060248
Change-Id: I7508b4775e2ab653426467ad8e289dc7b3e964a9